Add flag for loading D3 data. Add fallback to material system.

+++ src/Inferno/Graphics/Render.Level.cpp
@@ -472,6 +472,6 @@ namespace Inferno::Render {
     }
 
     int GetTransparentQueueSize() {
-        return _renderQueue.Transparent().size();
+        return (int)_renderQueue.Transparent().size();
     }
 }
